# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ANNEX API
  version: 0.1.0
paths:
  /health:
    get:
      tags:
      - system
      summary: Health
      description: Liveness probe — 200 whenever the process is serving requests.
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Health Health Get
  /health/ready:
    get:
      tags:
      - system
      summary: Readiness
      description: 'Readiness probe — reports each configured dependency''s status.


        Returns 200 when every registered check passes and 503 (degraded) as

        soon as one fails, with per-check detail for operators.'
      operationId: readiness_health_ready_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/v1/meta/version:
    get:
      tags:
      - system
      summary: Version
      description: Return the API name, version, and runtime environment.
      operationId: version_api_v1_meta_version_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Version Api V1 Meta Version Get
  /api/v1/users/me:
    get:
      tags:
      - users
      summary: Me
      description: 'Return the authenticated caller''s profile.


        Requires ``Authorization: Bearer <Firebase ID token>``. The user is

        hydrated into the database on first access (ADR-0005).'
      operationId: me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Me Api V1 Users Me Get
  /api/v1/analysis:
    post:
      tags:
      - analysis
      summary: Submit
      description: 'Submit text, a URL, or an image for claim analysis.


        Accepts anonymous submissions; anonymous analyses carry no owner and can

        only be read back by their holder while unauthenticated context exists.

        Returns 202 with the analysis so clients can poll for the report.'
      operationId: submit_api_v1_analysis_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnalysisRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Submit Api V1 Analysis Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - analysis
      summary: List Analyses
      description: List the caller's analyses, newest first (cursor-paginated).
      operationId: list_analyses_api_v1_analysis_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Analyses Api V1 Analysis Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analysis/{analysis_id}:
    get:
      tags:
      - analysis
      summary: Get Analysis
      description: Fetch one analysis with its report (owner only).
      operationId: get_analysis_api_v1_analysis__analysis_id__get
      parameters:
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Analysis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Analysis Api V1 Analysis  Analysis Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - analysis
      summary: Delete Analysis
      description: Delete an analysis (owner only).
      operationId: delete_analysis_api_v1_analysis__analysis_id__delete
      parameters:
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Analysis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Analysis Api V1 Analysis  Analysis Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/claims/{claim_id}:
    get:
      tags:
      - claims
      summary: Get Claim
      description: Fetch one claim with its verdict and rationale (owner only).
      operationId: get_claim_api_v1_claims__claim_id__get
      parameters:
      - name: claim_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Claim Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Claim Api V1 Claims  Claim Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/claims/{claim_id}/evidence:
    get:
      tags:
      - claims
      summary: Get Claim Evidence
      description: Fetch the evidence links supporting a claim's verdict (owner only).
      operationId: get_claim_evidence_api_v1_claims__claim_id__evidence_get
      parameters:
      - name: claim_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Claim Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Claim Evidence Api V1 Claims  Claim Id  Evidence
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sources/search:
    get:
      tags:
      - sources
      summary: Search Sources
      description: Search sources by domain or name (case-insensitive substring).
      operationId: search_sources_api_v1_sources_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 100
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Search Sources Api V1 Sources Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sources/{domain}:
    get:
      tags:
      - sources
      summary: Get Source
      description: Fetch one source profile with its credibility score.
      operationId: get_source_api_v1_sources__domain__get
      parameters:
      - name: domain
        in: path
        required: true
        schema:
          type: string
          title: Domain
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Source Api V1 Sources  Domain  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media:
    post:
      tags:
      - media
      summary: Submit Media
      description: Ingest an image for an owned analysis; returns the media record.
      operationId: submit_media_api_v1_media_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitMediaRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Submit Media Api V1 Media Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/{media_id}:
    get:
      tags:
      - media
      summary: Get Media
      description: Fetch one media record with OCR + forensics (owner only).
      operationId: get_media_api_v1_media__media_id__get
      parameters:
      - name: media_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Media Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Media Api V1 Media  Media Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/i18n/locales:
    get:
      tags:
      - i18n
      summary: List Locales
      description: Return every enabled locale with its fallback parent.
      operationId: list_locales_api_v1_i18n_locales_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response List Locales Api V1 I18N Locales Get
  /api/v1/i18n/bundles/{locale}:
    get:
      tags:
      - i18n
      summary: Get Bundle
      description: 'Return a versioned translation bundle resolved over the fallback
        chain.


        Conditional requests: pass the previously received ``version`` as a

        query parameter (or ``If-None-Match`` with the bundle ETag) to receive

        ``304 Not Modified`` while the bundle is unchanged.'
      operationId: get_bundle_api_v1_i18n_bundles__locale__get
      parameters:
      - name: locale
        in: path
        required: true
        schema:
          type: string
          title: Locale
      - name: version
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AnalysisInputType:
      type: string
      enum:
      - text
      - url
      - image
      title: AnalysisInputType
      description: The kind of content submitted for analysis.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    SubmitAnalysisRequest:
      properties:
        input_type:
          $ref: '#/components/schemas/AnalysisInputType'
          default: text
        text:
          anyOf:
          - type: string
            maxLength: 20000
            minLength: 1
          - type: 'null'
          title: Text
        url:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Url
        image:
          anyOf:
          - type: string
          - type: 'null'
          title: Image
        locale:
          type: string
          maxLength: 10
          minLength: 2
          title: Locale
          default: en
      type: object
      title: SubmitAnalysisRequest
      description: "Body for ``POST /analysis``.\n\nExactly one content field must\
        \ be present, matching ``input_type``:\n\n- ``text`` for text input,\n- ``url``\
        \ for URL input (fetched server-side, SSRF-guarded),\n- ``image`` for image\
        \ input — base64-encoded bytes or a ``data:`` URL;\n  validated at the API\
        \ boundary with a size cap, then OCR + forensics\n  (the MIME type is sniffed\
        \ from the bytes by the pipeline)."
    SubmitMediaRequest:
      properties:
        analysis_id:
          type: string
          format: uuid
          title: Analysis Id
        image:
          type: string
          title: Image
      type: object
      required:
      - analysis_id
      - image
      title: SubmitMediaRequest
      description: Body for ``POST /media``.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError